Rename package dir
