+++ src/tempor/datasources/mivdp/preproc/cohort/day_intervals_cohort_v1.py
